Introduce Abstract DictBased Printer
